# specs/001-feature-name-cv/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CV Optimization API
  description: API for CV upload, parsing, editing, and AI-enhanced optimization
  version: 1.0.0
  contact:
    name: CV Optimization Team
    email: support@cvoptimizer.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.cvoptimizer.com
    description: Production server
security:
  - BearerAuth: []
  - OAuth2: []
paths:
...
  /api/cvs:
    get:
      tags:
        - CV Management
      summary: List user's CVs
      description: Get all CVs for the authenticated user
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of CVs per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of CVs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - CV Management
      summary: Upload CV file
      description: Upload and parse a CV file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CV file (PDF, DOC, or DOCX)
              required:
                - file
      responses:
        '201':
          description: CV uploaded and parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVResponse'
        '400':
          description: Invalid file format or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cvs/{cv_id}:
    get:
      tags:
        - CV Management
      summary: Get CV details
      description: Get detailed information about a specific CV
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CV details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - CV Management
      summary: Update CV data
      description: Update CV parsed data
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CVUpdate'
      responses:
        '200':
          description: CV updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - CV Management
      summary: Delete CV
      description: Delete a CV and all associated data
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: CV deleted successfully
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Job Description Endpoints
  /api/cvs/{cv_id}/job-descriptions:
    get:
      tags:
        - Job Descriptions
      summary: Get job descriptions for CV
      description: Get all job descriptions associated with a CV
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job descriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDescriptionListResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Job Descriptions
      summary: Add job description
      description: Add a job description for CV optimization
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobDescriptionCreate'
      responses:
        '201':
          description: Job description added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDescriptionResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/job-descriptions/{jd_id}:
    delete:
      tags:
        - Job Descriptions
      summary: Delete job description
      description: Delete a job description
      parameters:
        - name: jd_id
          in: path
          required: true
          description: Job description identifier
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Job description deleted successfully
        '404':
          description: Job description not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # AI Features Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Email access
            profile: Profile access
  schemas:
...
    CVResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        original_filename:
          type: string
        file_size:
          type: integer
        file_type:
          type: string
        parsed_data:
          $ref: '#/components/schemas/CVParsedData'
        is_parsed:
          type: boolean
        parse_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CVUpdate:
      type: object
      properties:
        parsed_data:
          $ref: '#/components/schemas/CVParsedData'
    CVListResponse:
      type: object
      properties:
        cvs:
          type: array
          items:
            $ref: '#/components/schemas/CVResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
    CVParsedData:
      type: object
      properties:
        personal_info:
          $ref: '#/components/schemas/PersonalInfo'
        professional_summary:
          $ref: '#/components/schemas/ProfessionalSummary'
        work_experience:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        skills:
          $ref: '#/components/schemas/Skills'
        certifications:
          type: array
          items:
            $ref: '#/components/schemas/Certification'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        awards:
          type: array
          items:
            $ref: '#/components/schemas/Award'
        publications:
          type: array
          items:
            $ref: '#/components/schemas/Publication'
        volunteer_experience:
          type: array
          items:
            $ref: '#/components/schemas/VolunteerExperience'
    PersonalInfo:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        linkedin_url:
          type: string
        website_url:
          type: string
    ProfessionalSummary:
      type: object
      properties:
        content:
          type: string
        keywords:
          type: array
          items:
            type: string
    WorkExperience:
      type: object
      properties:
        company:
          type: string
        position:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        current:
          type: boolean
        description:
          type: string
        achievements:
          type: array
          items:
            type: string
        technologies:
          type: array
          items:
            type: string
    Education:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
        field_of_study:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        gpa:
          type: string
        honors:
          type: array
          items:
            type: string
    Skills:
      type: object
      properties:
        technical:
          type: array
          items:
            type: string
        soft:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
    Language:
      type: object
      properties:
        language:
          type: string
        proficiency:
          type: string
    Certification:
      type: object
      properties:
        name:
          type: string
        issuer:
          type: string
        date:
          type: string
        expiry_date:
          type: string
    Project:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        technologies:
          type: array
          items:
            type: string
        url:
          type: string
        start_date:
          type: string
        end_date:
          type: string
    Award:
      type: object
      properties:
        title:
          type: string
        issuer:
          type: string
        date:
          type: string
        description:
          type: string
    Publication:
      type: object
      properties:
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        journal:
          type: string
        date:
          type: string
        url:
          type: string
    VolunteerExperience:
      type: object
      properties:
        organization:
          type: string
        role:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        description:
          type: string

    # Job Description Schemas
    JobDescriptionCreate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Job description text
        source_url:
          type: string
          format: uri
          description: Original URL if scraped
        title:
          type: string
          description: Job title
        company:
          type: string
          description: Company name
        location:
          type: string
          description: Job location
    JobDescriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cv_id:
          type: string
          format: uuid
        content:
          type: string
        source_url:
          type: string
        title:
          type: string
        company:
          type: string
        location:
          type: string
        created_at:
          type: string
          format: date-time
    JobDescriptionListResponse:
      type: object
      properties:
        job_descriptions:
          type: array
          items:
            $ref: '#/components/schemas/JobDescriptionResponse'

    # AI Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        success:
          type: boolean
          default: false